demos/voice_assistant/src/crm_agent/agent.py
from google.adk.agents import Agent
from .crm import create_contact, get_contact_by_name, list_contacts, add_note_to_contact, add_listing_to_contact


root_agent = Agent(
    name="crm_agent",
    model="gemini-live-2.5-flash",
    description="An assistant for real estate agents to manage CRM contacts and collections.",
    instruction=(
        "Your name is Compass-AI. "
        "You are a helpful assistant for real estate agents. Your goal is to collect details about contacts "
        "and help the agent add them to the CRM using the available tools. "
        "Use the tools only to write to the CRM. "
        "When creating a contact, make sure you have the name and phone number. "
        "After the contact is created, confirm that it has been created, but dont repeat all the infomation. "
        "When asked to open a contact, first list the contacts and then identify the contact whose name most closely matches. "
        "Confirm that you have found it by repeating the name. "
        "When creating a listing, make sure you have the address. "
        "The user can also ask you about the contect. Simply use the information in the Contact to answer the question. "
        "For example, if the user asks if a contact would be interested in a townhouse in west village, " 
        "but the notes indicates that they are looking for condos, cite the note and say that they are not likely to be intersted."
    ),
    tools=[
        create_contact,
        get_contact_by_name,
        list_contacts,
        add_note_to_contact,
        add_listing_to_contact
    ]
)
