demos/voice_assistant/requirements.txt
google-adk
pydantic[email]
rich